refactor(cli): complete CLI modularization by removing old cli.py file
- Migrated all functionality from monolithic cli.py to dedicated modules
- Confirmed entry point in pyproject.toml correctly points to new structure
- Deleted redundant cli.py file (1699 lines) after confirming functionality
- Validated CLI functionality with version command test

The refactoring splits commands into logical modules:
- benchmark_commands.py
- data_commands.py
- text_commands.py
- index_commands.py
- query_commands.py
- utils.py
- __init__.py (main entry point)

This improves code maintainability, readability, and makes future extensions easier.
